backend:Gestione errori controller + service

+++ backend/src/main/java/com/istabasic/backend/service/ProfileService.java
@@ -1,14 +1,13 @@
 package com.istabasic.backend.service;
 
-import com.istabasic.backend.controller.CommentController;
-import com.istabasic.backend.model.Follow;
+import com.istabasic.backend.common.util.ErrorHandler;
+
 import com.istabasic.backend.model.Post;
 import com.istabasic.backend.model.Profile;
 import com.istabasic.backend.repository.FollowRepository;
 import com.istabasic.backend.repository.PostRepository;
 import com.istabasic.backend.repository.ProfileRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.apache.logging.log4j.LogManager;
@@ -30,29 +29,44 @@ public class ProfileService {
     static final Logger logger = LogManager.getLogger(ProfileService.class.getName());
 
     // C
-    public Profile save(Profile Profile) {
-        return ProfileRepository.save(Profile);
+    public Profile save(Profile profile) {
+        if (profile != null) {
+            return ProfileRepository.save(profile);
+        } else {
+            throw new ErrorHandler(400, "null");
+        }
     }
 
     // R
-    public Optional<Profile> findById(Long id) {
-        return ProfileRepository.findById(id);
+    public Profile findById(Long id) {
+        if (id != null) {
+            Optional<Profile> optional = ProfileRepository.findById(id);
+            if (optional.isPresent()) {
+                return optional.get();
+            } else {
+                throw new ErrorHandler(404, "Comment not found");
+            }
+        } else {
+            throw new ErrorHandler(400, "Id null");
+        }
     }
 
     public Profile findByProfilename(String name) {
-        Profile Profile = null;
         Optional<Profile> optional = ProfileRepository.findByProfilename(name);
         if (optional.isPresent()) {
-            Profile = optional.get();
+            return optional.get();
         } else {
-            logger.warn("Profile not found");
+            throw new ErrorHandler(404, "Profile not found");
         }
-        return Profile;
-
     }
 
     public Page<Profile> findByUserId(String userId, Pageable pageable) {
-        return ProfileRepository.findByUserId(userId, pageable);
+        if (userId == null) {
+            return ProfileRepository.findByUserId(userId, pageable);
+        } else {
+            throw new ErrorHandler(400, "userId=null");
+        }
+
     }
 
     public Page<Post> getPostsByFollow(String profileId, Pageable pageable) {
@@ -73,21 +87,32 @@ public class ProfileService {
 
     // U
     public Profile update(Long id, Profile Profile) {
-        Optional<Profile> ProfileResult = ProfileRepository.findById(id);
-
+        Optional<Profile> ProfileResult = null;
+        if (id != null) {
+            ProfileResult = ProfileRepository.findById(id);
+        } else {
+            throw new ErrorHandler(400, "Profile id is null");
+        }
         if (ProfileResult.isPresent()) {
             Profile ProfileUpdate = ProfileResult.get();
-
-            ProfileRepository.save(ProfileUpdate);
-            return ProfileUpdate;
+            if (ProfileUpdate != null) {
+                ProfileRepository.save(ProfileUpdate);
+                return ProfileUpdate;
+            } else {
+                throw new ErrorHandler(404, "Profile not found");
+            }
         } else {
-            throw new Error("Profile not found"); // TODO:GESTIONE ERRORI
+            throw new ErrorHandler(404, "Profile not found");
         }
 
     }
 
     // D
     public void delete(Long id) {
-        ProfileRepository.deleteById(id);
+        if (id != null) {
+            ProfileRepository.deleteById(id);
+        } else {
+            throw new ErrorHandler(404, "Comment not found");
+        }
     }
 }
